Fix view transitions creating an unescapable UI state
vt's cause input blocking
Refuse to play < 600ms, as it would only havea window of about 300ms (50% of the time) where clicks actually land on an element.

Add listener to catch clicks that land on the view transition
Seems that this is the only way clicks land directly onto the root, so should be safe.