Lock history to avoid unstable pin scenarios, don't skip over pinned state
This commit should avoid most (perhaps all) of the scenarios where the pin system got in an invalid state.
1: Pin is in the future, in removed section of the timeline. This resulted in the pins not properly stored, causing different changes when you reload the timeline. Any such change is not processed anymore. This provides the ability to easily protect the timeline without using "demo mode".
2: Pin is not on the latest theme edit. With how it was implemented, it would create a state that does not contain the entries in between. However as the page was reloaded they would again be included.

Also, when fast navigating, don't skip over entries wtih a pin if that entry is where it's pinned.

Somewhat related, also add a equality check to avoid dispatching using the useGlobalState hook. This hook might be misbehaving due to the architecture and recent changes to React.