Better style updates, directly update frames sheets
Also comment some unneeded experiments.

It is now able to update only the properties that are necessary in all cases,
without needing to throw away the sheet like it did before.

It's unclear whether it's better to update all frames together with the UI,
or render the UI first, as what happend when sending a message previously.

Possibly the styles can be fast enough that this can happen in one go,
but for now doing it in one pass can cause long INP